+++ Mods ChoGGi/Spice Harvester/metadata.lua
@@ -1,12 +1,16 @@
 return PlaceObj("ModDef", {
-  "title", "Spice Harvester v0.2",
-  "version", 2,
-  "saved", 1529323200,
+  "title", "Spice Harvester v0.3",
+  "version", 3,
+  "saved", 1535457600,
 	"image", "Preview.png",
   "tags", "Buildings",
   "id", "ChoGGi_SpiceHarvester",
   "author", "ChoGGi",
-  "code", {"Script.lua"},
+  "code", {
+		"Code/1 CommonFunctions.lua",
+		"Code/2 OnMsgs.lua",
+		"Code/SpiceHarvester.lua",
+	},
 	"steam_id", "1416040484",
 	"lua_revision", LuaRevision,
   "description", [[It doesn't do much; but move around and make thumping sounds at the moment.